refactor chat page to use ChatSidebar component
